Pins should not be dragged along when changing locations.
Instead, the old pins should disappear (since they are no longer relevant), and the new pins should appear when the new theater list is loaded.
Also gets around the map view graphical glitch that will cause pins to be "dragged along" as the map view scrolls to a new location.